page access control functionality
