+++ netmanager-app/app/(authenticated)/layout.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import "../globals.css";
-import Layout from "../components/layout";
+import Layout from "../../components/layout";
 
 export default function AuthenticatedLayout({
   children,
